@@ -1,56 +1,36 @@
-import { Component, ViewChild, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { IonSegment } from '@ionic/angular';
 import { NoticiasService } from '../../services/noticias.service';
-import { Article } from '../../interfaces/interfaces';
+import { Article } from '../interfaces/interfaces';
 
 @Component({
   selector: 'app-tab2',
   templateUrl: 'tab2.page.html',
   styleUrls: ['tab2.page.scss']
 })
-export class Tab2Page implements OnInit {
+export class Tab2Page  implements OnInit{
 
-  @ViewChild(IonSegment) segment: IonSegment;
+  //@ViewChild(IonSegment, {static: false}) segment: IonSegment;
 
-  categorias = ['business', 'entertainment', 'general', 'health', 'science', 'sports', 'technology', ];
+  categorias = ['business', 'entertainment', 'general', 'health', 'science', 'sports', 'technology'];
   noticias: Article[] = [];
 
+  constructor(private noticiasService: NoticiasService) {}
 
-  constructor( private noticiasService: NoticiasService ) {
-
+  ngOnInit(){
+    this.cargarNoticias(this.categorias[0]);
   }
 
-
-  ngOnInit() {
-    this.segment.value = this.categorias[0];
-
-    this.cargarNoticias( this.categorias[0] );
-  }
-
-  cambioCategoria( event ) {
-
+  cambioCategoria( event ){
     this.noticias = [];
-
-    this.cargarNoticias( event.detail.value );
-
+    this.cargarNoticias(event.detail.value);
   }
 
-  cargarNoticias( categoria: string, event? ) {
-
-    this.noticiasService.getTopHeadlinesCategoria( categoria )
-          .subscribe( resp => {
-            // console.log(resp);
-            this.noticias.push( ...resp.articles );
-
-            if ( event ) {
-              event.target.complete();
-            }
-          });
-  }
-
-  loadData( event ) {
-
-    this.cargarNoticias( this.segment.value, event );
+  cargarNoticias(categoria: string){
+    this.noticiasService.getTopHeadLinesCategoria(categoria)
+    .subscribe( resp => {
+      this.noticias.push(...resp.articles);
+    });
 
   }
 
