@@ -6,13 +6,13 @@ var WebView = {
     if (!url) {
       return url;
     }
-    if (url.startsWith('/')) {
+    if (url.indexOf('/')===0) {
       return window.WEBVIEW_SERVER_URL + '/_app_file_' + url;
     }
-    if (url.startsWith('file://')) {
+    if (url.indexOf('file://')===0) {
       return window.WEBVIEW_SERVER_URL + url.replace('file://', '/_app_file_');
     }
-    if (url.startsWith('content://')) {
+    if (url.indexOf('content://')===0) {
       return window.WEBVIEW_SERVER_URL + url.replace('content:/', '/_app_content_');
     }
     return url;
@@ -29,4 +29,5 @@ var WebView = {
 }
 
 module.exports = WebView;
+
 });
