@@ -3,7 +3,6 @@ var argscheck = require('cordova/argscheck'),
     exec = require('cordova/exec'),
     channel = require('cordova/channel');
 
-
 var Keyboard = function () {};
 
 Keyboard.fireOnShow = function (height) {
@@ -39,27 +38,31 @@ Keyboard.fireOnShowing = function (height) {
 };
 
 Keyboard.hideFormAccessoryBar = Keyboard.hideKeyboardAccessoryBar = function (hide) {
-    exec(null, null, "Keyboard", "hideKeyboardAccessoryBar", [hide]);
+    console.warn("Keyboard.hideKeyboardAccessoryBar() not supported in Android");
 };
 
 Keyboard.hide = function () {
-    exec(null, null, "Keyboard", "hide", []);
+    exec(null, null, "CDVIonicKeyboard", "hide", []);
 };
 
 Keyboard.show = function () {
-    exec(null, null, "Keyboard", "show", []);
+    exec(null, null, "CDVIonicKeyboard", "show", []);
 };
 
 Keyboard.disableScroll = function (disable) {
-    console.warn("Keyboard.disableScroll() was removed");
+    console.warn("Keyboard.disableScroll() not supported in Android");
 };
 
 Keyboard.setResizeMode = function (mode) {
     console.warn("Keyboard.setResizeMode() not supported in Android");
 }
 
+Keyboard.setKeyboardStyle = function(style) {
+    console.warn("Keyboard.setKeyboardStyle() not supported in Android");
+};
+
 channel.onCordovaReady.subscribe(function () {
-    exec(success, null, 'Keyboard', 'init', []);
+    exec(success, null, 'CDVIonicKeyboard', 'init', []);
 
     function success(msg) {
         var action = msg.charAt(0);
