@@ -1,5 +1,6 @@
 #import "IONAssetHandler.h"
 #import <MobileCoreServices/MobileCoreServices.h>
+#import "CDVWKWebViewEngine.h"
 
 @implementation IONAssetHandler
 
@@ -7,17 +8,27 @@
     self.basePath = assetPath;
 }
 
+- (instancetype)initWithBasePath:(NSString *)basePath andScheme:(NSString *)scheme {
+    self = [super init];
+    if (self) {
+        _basePath = basePath;
+        _scheme = scheme;
+    }
+    return self;
+}
+
 - (void)webView:(WKWebView *)webView startURLSchemeTask:(id <WKURLSchemeTask>)urlSchemeTask
-API_AVAILABLE(ios(11.0)){
+{
     NSString * startPath = @"";
     NSURL * url = urlSchemeTask.request.URL;
     NSString * stringToLoad = url.path;
     NSString * scheme = url.scheme;
-    if ([scheme isEqualToString:@"ionic"]) {
+
+    if ([scheme isEqualToString:self.scheme]) {
         if ([stringToLoad hasPrefix:@"/_app_file_"]) {
             startPath = [stringToLoad stringByReplacingOccurrencesOfString:@"/_app_file_" withString:@""];
         } else {
-            startPath = self.basePath;
+            startPath = self.basePath ? self.basePath : @"";
             if ([stringToLoad isEqualToString:@""] || [url.pathExtension isEqualToString:@""]) {
                 startPath = [startPath stringByAppendingString:@"/index.html"];
             } else {
@@ -25,8 +36,14 @@ API_AVAILABLE(ios(11.0)){
             }
         }
     }
-    
-    NSData * data = [[NSData alloc] initWithContentsOfFile:startPath];
+    NSError * fileError = nil;
+    NSData * data = nil;
+    if ([self isMediaExtension:url.pathExtension]) {
+        data = [NSData dataWithContentsOfFile:startPath options:NSDataReadingMappedIfSafe error:&fileError];
+    }
+    if (!data || fileError) {
+        data =  [[NSData alloc] initWithContentsOfFile:startPath];
+    }
     NSInteger statusCode = 200;
     if (!data) {
         statusCode = 404;
@@ -47,7 +64,8 @@ API_AVAILABLE(ios(11.0)){
 
 }
 
-- (void)webView:(nonnull WKWebView *)webView stopURLSchemeTask:(nonnull id<WKURLSchemeTask>)urlSchemeTask  API_AVAILABLE(ios(11.0)){
+- (void)webView:(nonnull WKWebView *)webView stopURLSchemeTask:(nonnull id<WKURLSchemeTask>)urlSchemeTask
+{
     NSLog(@"stop");
 }
 
