@@ -1,133 +1,17 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Article } from '../../interfaces/interfaces';
-
-import { InAppBrowser } from '@ionic-native/in-app-browser/ngx';
-import { ActionSheetController, Platform } from '@ionic/angular';
-import { SocialSharing } from '@ionic-native/social-sharing/ngx';
-import { DataLocalService } from '../../services/data-local.service';
+import { Component, Input, OnInit } from '@angular/core';
+import { Article } from '../../pages/interfaces/interfaces';
 
 @Component({
   selector: 'app-noticia',
   templateUrl: './noticia.component.html',
-  styleUrls: ['./noticia.component.scss']
+  styleUrls: ['./noticia.component.scss'],
 })
 export class NoticiaComponent implements OnInit {
-
   @Input() noticia: Article;
   @Input() indice: number;
-  @Input() enFavoritos;
 
-  constructor( private iab: InAppBrowser,
-               private actionSheetCtrl: ActionSheetController,
-               private socialSharing: SocialSharing,
-               private datalocalService: DataLocalService,
-               private platform: Platform ) { }
+  constructor() { }
 
-  ngOnInit() {
-
-    console.log('Favoritos', this.enFavoritos );
-
-  }
-
-  abrirNoticia() {
-    // console.log('Noticia', );
-    const browser = this.iab.create(this.noticia.url, '_system');
-
-  }
-
-  async lanzarMenu() {
-
-    let guardarBorrarBtn;
-
-    if ( this.enFavoritos ) {
-
-      guardarBorrarBtn = {
-        text: 'Borrar Favorito',
-        icon: 'trash',
-        cssClass: 'action-dark',
-        handler: () => {
-          console.log('Borrar de favorito');
-          this.datalocalService.borrarNoticia( this.noticia );
-        }
-      };
-
-    } else {
-
-      guardarBorrarBtn = {
-        text: 'Favorito',
-        icon: 'star',
-        cssClass: 'action-dark',
-        handler: () => {
-          console.log('Favorito');
-          this.datalocalService.guardarNoticia( this.noticia );
-        }
-      };
-
-    }
-
-
-
-    const actionSheet = await this.actionSheetCtrl.create({
-      buttons: [
-        {
-          text: 'Compartir',
-          icon: 'share',
-          cssClass: 'action-dark',
-          handler: () => {
-            console.log('Share clicked');
-
-            this.compartirNoticia();
-
-        }
-      },
-      guardarBorrarBtn,
-      {
-        text: 'Cancelar',
-        icon: 'close',
-        role: 'cancel',
-        cssClass: 'action-dark',
-        handler: () => {
-          console.log('Cancel clicked');
-        }
-      }]
-    });
-
-    await actionSheet.present();
-
-  }
-
-
-  compartirNoticia() {
-
-
-    if ( this.platform.is('cordova') ) {
-      this.socialSharing.share(
-        this.noticia.title,
-        this.noticia.source.name,
-        '',
-        this.noticia.url
-      );
-    } else {
-
-      if (navigator['share'] ) {
-
-        navigator['share']({
-            title: this.noticia.title,
-            text: this.noticia.description,
-            url: this.noticia.url,
-        })
-          .then(() => console.log('Successful share'))
-          .catch((error) => console.log('Error sharing', error));
-      } else {
-        console.log('No se pudo compartir porque no se soporta');
-      }
-
-    }
-
-
-
-
-  }
-
+  ngOnInit() {}
 
 }
