@@ -1,20 +1,17 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Article } from '../../interfaces/interfaces';
+import { Article } from '../../pages/interfaces/interfaces';
 
 @Component({
   selector: 'app-noticias',
   templateUrl: './noticias.component.html',
-  styleUrls: ['./noticias.component.scss']
+  styleUrls: ['./noticias.component.scss'],
 })
 export class NoticiasComponent implements OnInit {
 
   @Input() noticias: Article[] = [];
-  @Input() enFavoritos = false;
-
 
   constructor() { }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
 }
