@@ -4,15 +4,19 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Tab2Page } from './tab2.page';
-import { ComponentsModule } from '../../components/components.module';
+import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
+
+import { Tab2PageRoutingModule } from './tab2-routing.module';
+import { ComponentsModule } from 'src/app/components/components.module';
 
 @NgModule({
   imports: [
     IonicModule,
     CommonModule,
     FormsModule,
+    ExploreContainerComponentModule,
     ComponentsModule,
-    RouterModule.forChild([{ path: '', component: Tab2Page }])
+    Tab2PageRoutingModule
   ],
   declarations: [Tab2Page]
 })
