@@ -4,15 +4,18 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { Tab3Page } from './tab3.page';
-import { ComponentsModule } from '../../components/components.module';
+import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
+
+import { Tab3PageRoutingModule } from './tab3-routing.module'
 
 @NgModule({
   imports: [
     IonicModule,
     CommonModule,
     FormsModule,
-    ComponentsModule,
-    RouterModule.forChild([{ path: '', component: Tab3Page }])
+    ExploreContainerComponentModule,
+    RouterModule.forChild([{ path: '', component: Tab3Page }]),
+    Tab3PageRoutingModule,
   ],
   declarations: [Tab3Page]
 })
