@@ -1,5 +1,5 @@
 export const environment = {
   production: true,
   apiKey: 'd268e175440744edbba4a0b3bffecfbe',
-  apiUlr: 'https://newsapi.org/v2'
+  apiUrl: 'https://newsapi.org/v2'
 };
